% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatePlotParam.R
\name{updatePlotParam}
\alias{updatePlotParam}
\title{updatePlotParam}
\usage{
updatePlotParam(obj, metadata = NULL, value = NULL)
}
\arguments{
\item{obj}{a list object. An object which contain all parameters
and metadata of gfplots}

\item{metadata}{a character. Target to modify metadata}

\item{value}{value of the metadata}
}
\value{
an object
}
\description{
updatePlotParam
}
\author{
Yuting Dai
}
