% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfplots.R
\name{gfplots}
\alias{gfplots}
\title{main function of gfplots packages}
\usage{
gfplots(x, db.type = "ensembl", output.directory = ".",
  plot.type = "mutation", version = "GRCh37", plot.condig = NULL)
}
\arguments{
\item{x}{data.frame. The basic information that need to be plotted
using gfplots. If \param{plot.type} is "mutation", then the
data.frame show only have}

\item{db.type}{string.}

\item{output.directory}{string. The absolute path of your
output dirctory which contains your report.}

\item{plot.type}{string. "mutation" or "fusion". Different plot.type
corresponds different plot style.}
}
\value{

}
\description{
\code{gfplots} provides an
}
\details{
gfplots packages
}
\examples{

}
\author{
Yuting Dai
}
